mesh/pyproject.toml
[project]
name = "mesh"
version = "1.0.0"
description = "Mesh network setup tools (Headscale + Syncthing)"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.15",
    "rich>=13.0",
    "httpx>=0.28",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
mesh = "mesh.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mesh"]

[tool.ty.environment]
python-version = "3.13"

[dependency-groups]
dev = [
    "pyinstaller>=6.0",
    "pytest>=8.0",
    "ruff>=0.8",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
